requirements.txt
Flask
requests
cohere
groq
google-generativeai
python-dotenv